JWT 인증 로직 추가 * 인증, 인가 로직 예시 추가 * 공공데이터 DB 업로드 로직 수정

+++ server/eventHomePage/src/main/java/com/ict/eventHomePage/events/repository/EventsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventsRepository extends JpaRepository<Events, Long> {
+public interface EventsRepository extends JpaRepository<Events, Integer> {
 }
